@@ -1,50 +1,25 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import { PropTypes } from 'prop-types';
-import { connect } from 'react-redux';
+import React from 'react';
+import {Link} from 'react-router-dom';
 
-class Landing extends Component {
-  componentDidMount() {
-    if (this.props.auth.isAuthenticated) {
-      this.props.history.push('/dashboard');
-    }
-  }
 
-  render() {
-    return (
-      <div className="landing">
-        <div className="dark-overlay landing-inner text-light">
-          <div className="container">
-            <div className="row">
-              <div className="col-md-12 text-center">
-                <h1 className="display-3 mb-4">Developer Connector</h1>
-                <p className="lead">
-                  {' '}
-                  Create a developer profile/portfolio, share posts and get help
-                  from other developers
-                </p>
-                <hr />
-                <Link to="/register" className="btn btn-lg btn-info mr-2">
-                  Sign Up
-                </Link>
-                <Link to="/login" className="btn btn-lg btn-light">
-                  Login
-                </Link>
-              </div>
-            </div>
-          </div>
+const Landing = () => {
+  return (
+    <section className="landing">
+    <div className="dark-overlay">
+      <div className="landing-inner">
+        <h1 className="x-large">Developer Connector</h1>
+        <p className="lead">
+          Create a developer profile/portfolio, share posts and get help from
+          other developers
+        </p>
+        <div className="buttons">
+          <Link to="/Register" className="btn btn-primary">Sign Up</Link>
+          <Link to="/Login" className="btn btn-light">Login</Link>
         </div>
       </div>
-    );
-  }
+    </div>
+  </section>
+  )
 }
 
-Landing.propTypes = {
-  auth: PropTypes.object.isRequired
-};
-
-const mapStateToProps = state => ({
-  auth: state.auth
-});
-
-export default connect(mapStateToProps)(Landing);
+export default Landing
