@@ -1,17 +1,10 @@
-import { createStore, applyMiddleware, compose } from 'redux';
+import {createStore, applyMiddleware} from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers';
-
-const initialState = {};
-
-const middleware = [thunk];
-
+import rootReducer from './Reducer';
+const middleware =[thunk];
 const store = createStore(
-  rootReducer,
-  initialState,
-  compose(
+    rootReducer,
+    {},
     applyMiddleware(...middleware)
-  )
 );
-
 export default store;
